+++ app/code/community/Pubble/Messenger/Helper/Data.php
@@ -5,8 +5,8 @@
  * @category  Pubble
  * @package   Pubble_Messenger
  * @author    Pubble <ross@pubble.io>
- * @copyright 2015 Pubble (http://www.pubble.io)
- * @version   1.0.0
+ * @copyright 2016 Pubble (http://www.pubble.io)
+ * @version   1.1.0
  */
 
 /**
@@ -18,14 +18,6 @@
  */
 class Pubble_Messenger_Helper_Data extends Mage_Core_Helper_Abstract
 {
-    /**#@+
-     * Configuration Method Types
-     * @var int
-     */
-    const METHOD_CREDENTIALS = 1;
-    const METHOD_COPYPASTE   = 2;
-    /**#@-*/
-    
     /**
      * Determine whether the module is enabled via admin configuration.
      *
@@ -34,36 +26,6 @@ class Pubble_Messenger_Helper_Data extends Mage_Core_Helper_Abstract
     public function isEnabled()
     {
         return Mage::getStoreConfigFlag('pubble/messenger/enabled');
-    }
-    
-    /**
-     * Get the method configuration type.
-     *
-     * @return int
-     */
-    public function getMethod()
-    {
-        return (int) Mage::getStoreConfig('pubble/messenger/method');
-    }
-    
-    /**
-     * Determine if the method type is 'credentials'.
-     *
-     * @return bool
-     */
-    public function isMethodCredentials()
-    {
-        return ($this->getMethod() === static::METHOD_CREDENTIALS);
-    }
-    
-    /**
-     * Determine if the method type is 'copy and paste'.
-     *
-     * @return bool
-     */
-    public function isMethodCopyPaste()
-    {
-        return ($this->getMethod() === static::METHOD_COPYPASTE);
     }
     
     /**
@@ -84,15 +46,5 @@ class Pubble_Messenger_Helper_Data extends Mage_Core_Helper_Abstract
     public function getIdentifier()
     {
         return Mage::getStoreConfig('pubble/messenger/identifier');
-    }
-    
-    /**
-     * Get the copy and pasted code from Pubble saved in the configuration.
-     *
-     * @return int
-     */
-    public function getCode()
-    {
-        return Mage::getStoreConfig('pubble/messenger/code');
     }
 }