+++ app/code/community/Pubble/Messenger/data/pubble_messenger_setup/data-install-1.0.0.php
@@ -1,0 +1,3 @@
+<?php
+
+// No data to update