+++ app/code/community/Pubble/Messenger/data/pubble_messenger_setup/data-upgrade-1.0.0-1.1.0.php
@@ -1,0 +1,10 @@
+<?php
+
+$paths = array('pubble/messenger/method', 'pubble/messenger/code');
+foreach ($paths as $path) {
+    Mage::getConfig()->deleteConfig($path);
+}
+
+if (2 === Mage::getStoreConfigFlag('pubble/messenger/method')) {
+    Mage::getConfig()->saveConfig('pubble/messenger/enabled', 0);
+}